Relocate copy_args routine-definition and declaration to interface.
