+++ acados/dense_qp/dense_qp_hpipm.c
@@ -66,20 +66,6 @@ void *dense_qp_hpipm_assign_args(dense_qp_dims *dims, void *raw_memory)
 
 
 
-void *dense_qp_hpipm_copy_args(dense_qp_dims *dims, void *raw_memory, void *source_)
-{
-    dense_qp_hpipm_args *source = (dense_qp_hpipm_args *)source_;
-    dense_qp_hpipm_args *dest;
-
-    dest = (dense_qp_hpipm_args *) dense_qp_hpipm_assign_args(dims, raw_memory);
-
-    *dest->hpipm_args = *source->hpipm_args;
-
-    return (void *)dest;
-}
-
-
-
 void dense_qp_hpipm_initialize_default_args(void *args_)
 {
     dense_qp_hpipm_args *args = (dense_qp_hpipm_args *)args_;
