+++ acados/dense_qp/dense_qp_hpipm.h
@@ -48,8 +48,6 @@ void dense_qp_hpipm_initialize_default_args(void *args_);
 //
 void *dense_qp_hpipm_assign_args(dense_qp_dims *dims, void *raw_memory);
 //
-void *dense_qp_hpipm_copy_args(dense_qp_dims *dims, void *raw_memory, void *source_);
-//
 int dense_qp_hpipm_calculate_memory_size(dense_qp_dims *dims, void *args_);
 //
 void *dense_qp_hpipm_assign_memory(dense_qp_dims *dims, void *args_, void *raw_memory);
