+++ acados/dense_qp/dense_qp_qpoases.h
@@ -99,8 +99,6 @@ int dense_qp_qpoases_calculate_args_size(dense_qp_dims *dims);
 //
 void *dense_qp_qpoases_assign_args(dense_qp_dims *dims, void *raw_memory);
 //
-void *dense_qp_qpoases_copy_args(dense_qp_dims *dims, void *raw_memory, void *source_);
-//
 void dense_qp_qpoases_initialize_default_args(void *args_);
 //
 int dense_qp_qpoases_calculate_memory_size(dense_qp_dims *dims, void *args_);
