+++ acados/ocp_qp/ocp_qp_full_condensing.c
@@ -64,17 +64,6 @@ ocp_qp_full_condensing_args *ocp_qp_full_condensing_assign_args(ocp_qp_dims *dim
 
 
 
-ocp_qp_full_condensing_args *ocp_qp_full_condensing_copy_args(ocp_qp_dims *dims, void *raw_memory, ocp_qp_full_condensing_args *source_)
-{
-    ocp_qp_full_condensing_args *dest;
-
-    dest = ocp_qp_full_condensing_assign_args(dims, raw_memory);
-
-    return dest;
-}
-
-
-
 int ocp_qp_full_condensing_calculate_memory_size(ocp_qp_dims *dims, ocp_qp_full_condensing_args *args)
 {
     int size = 0;
