+++ acados/ocp_qp/ocp_qp_hpmpc.h
@@ -63,8 +63,6 @@ int ocp_qp_hpmpc_calculate_args_size(ocp_qp_dims *dims);
 //
 void *ocp_qp_hpmpc_assign_args(ocp_qp_dims *dims, void *raw_memory);
 //
-void *ocp_qp_hpmpc_copy_args(ocp_qp_dims *dims, void *raw_memory, void *source_);
-//
 void ocp_qp_hpmpc_initialize_default_args(void *args_);
 //
 int ocp_qp_hpmpc_calculate_memory_size(ocp_qp_dims *dims, void *args_);
