+++ acados/ocp_qp/ocp_qp_partial_condensing.h
@@ -47,8 +47,6 @@ int ocp_qp_partial_condensing_calculate_args_size(ocp_qp_dims *dims);
 //
 ocp_qp_partial_condensing_args *ocp_qp_partial_condensing_assign_args(ocp_qp_dims *dims, void *raw_memory);
 //
-ocp_qp_partial_condensing_args *ocp_qp_partial_condensing_copy_args(ocp_qp_dims *dims, void *raw_memory, ocp_qp_partial_condensing_args *source_);
-//
 void ocp_qp_partial_condensing_initialize_default_args(ocp_qp_partial_condensing_args *args);
 //
 int ocp_qp_partial_condensing_calculate_memory_size(ocp_qp_dims *dims, ocp_qp_partial_condensing_args *args);
