+++ interfaces/acados_c/ocp_qp/ocp_qp_full_condensing.h
@@ -24,6 +24,11 @@
 extern "C" {
 #endif
 
+#include <acados/ocp_qp_ocp_qp_full_condensing.h>
+
+//
+ocp_qp_full_condensing_args *ocp_qp_full_condensing_copy_args(ocp_qp_dims *dims, void *raw_memory, ocp_qp_full_condensing_args *source_);
+
 #ifdef __cplusplus
 } /* extern "C" */
 #endif
