+++ interfaces/acados_c/ocp_qp/ocp_qp_hpipm.c
@@ -18,3 +18,17 @@
  */
 
 #include "acados_c/ocp_qp/ocp_qp_hpipm.h"
+
+
+
+void *ocp_qp_hpipm_copy_args(ocp_qp_dims *dims, void *raw_memory, void *source_)
+{
+    ocp_qp_hpipm_args *source = (ocp_qp_hpipm_args *) source_;
+    ocp_qp_hpipm_args *dest;
+
+    dest = (ocp_qp_hpipm_args *) ocp_qp_hpipm_assign_args(dims, raw_memory);
+
+    *dest->hpipm_args = *source->hpipm_args;
+
+    return (void *)dest;
+}
