+++ interfaces/acados_c/ocp_qp/ocp_qp_partial_condensing.c
@@ -18,3 +18,18 @@
  */
 
 #include "acados_c/ocp_qp/ocp_qp_partial_condensing.h"
+
+
+
+ocp_qp_partial_condensing_args *ocp_qp_partial_condensing_copy_args(ocp_qp_dims *dims, void *raw_memory, ocp_qp_partial_condensing_args *source_)
+{
+    ocp_qp_partial_condensing_args *source = (ocp_qp_partial_condensing_args *) source_;
+    ocp_qp_partial_condensing_args *dest;
+
+    dest = ocp_qp_partial_condensing_assign_args(dims, raw_memory);
+
+    dest->N2 = source->N2;
+    dest->pcond_dims->N = source->pcond_dims->N;
+
+    return dest;
+}
